# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "commerce-system API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/products": {
            "get": {
                "description": "상품 목록을 필터링하여 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "상품 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "카테고리",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "상품명",
                        "name": "product_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "상품 목록",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.ProductResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "조회 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "새로운 상품을 등록합니다. (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "상품 생성",
                "parameters": [
                    {
                        "description": "상품 정보",
                        "name": "productRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "생성 성공",
                        "schema": {
                            "$ref": "#/definitions/response.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "생성 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/{product_number}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "상품을 삭제합니다. (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "상품 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "상품 번호",
                        "name": "product_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 상품 번호",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "삭제 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/{product_number}/stock": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "상품의 재고 수량을 수정합니다. (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "재고 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "상품 번호",
                        "name": "product_number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "재고 정보",
                        "name": "stockRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "수정 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.CreateProductRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "product_number": {
                    "type": "string"
                },
                "stock_quantity": {
                    "type": "integer"
                }
            }
        },
...
        "request.UpdateStockRequest": {
            "type": "object",
            "properties": {
                "stock_quantity": {
                    "type": "integer"
                }
            }
        },
...
        "response.ProductResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "product_number": {
                    "type": "string"
                },
                "stock_quantity": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "A commerce-system service API in Go using Gin framework",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}